backend/requirements.txt
fastapi
uvicorn
bandit
semgrep
radon
lizard
requests
python-dotenv
pymongo
mysql-connector-python
